Relocate a few helpers from qtscriptfuncs
