+++ src/hci.h
@@ -41,7 +41,6 @@ struct PROXIMITY_DISPLAY;
 struct STRUCTURE;
 struct W_SCREEN;
 struct iIMDShape;
-class QScriptEngine;
 
 enum  				  // Reticule button indecies.
 {
